Revert "fixed 32bits and 64bits problem of trie dict"
This reverts commit a9ea950c1ccb8b05ccc788ec39ac02d6434437bf.

+++ source/trie/double_array.cxx
@@ -171,7 +171,7 @@ int DoubleArray::search(const char *key)
 	return s?_base(s):0;
 }
 
-void DoubleArray::_explore(on_explore_finish_t cb, void *arg, int s, int off)
+void DoubleArray::_explore(on_explore_finish_t cb, void *arg, int s, size_t off)
 {
 	int *p, key[alphabet_size];
 	int t;
